changes for g7

+++ sweet_cash/api/models/transaction.py
@@ -1,19 +1,29 @@
 
+import enum
 from datetime import datetime
 
 from db import db
+from api.models.base import BaseModel
 
 
-class TransactionModel(db.Model):
+class TransactionType(enum.Enum):
+    INCOME = "Income"
+    EXPENSE = "Expense"
+
+    @classmethod
+    def has_value(cls, value):
+        return value in cls._value2member_map_
+
+
+class TransactionModel(BaseModel):
     __tablename__ = 'transactions'
     __table_args__ = {'extend_existing': True}
     id = db.Column(db.Integer, primary_key=True)
-    type = db.Column(db.Integer, nullable=False)
+    type = db.Column(db.Enum(TransactionType), nullable=False)
     user_id = db.Column(db.Integer, index=True, nullable=False)
     category = db.Column(db.Integer, nullable=False)
     amount = db.Column(db.Float, nullable=False)
     transaction_date = db.Column(db.DateTime, nullable=False)
-    created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow())
     private = db.Column(db.Boolean, nullable=False, default=False)
     description = db.Column(db.String(250), nullable=True)
 
@@ -26,18 +36,8 @@ class TransactionModel(db.Model):
         self.private = kwargs.get('private')
         self.description = kwargs.get('description')
 
-    def get_id(self):
-        return self.id
-
-    def get_created_at(self):
-        return self.created_at
-
-    def create(self):
-        db.session.add(self)
-        db.session.commit()
-
     def update(self, **kwargs):
-        self.type = kwargs.get('type_id')
+        self.type = kwargs.get('type')
         self.category = kwargs.get('category_id')
         self.amount = kwargs.get('amount')
         self.transaction_date = kwargs.get('transaction_date')
@@ -61,6 +61,7 @@ class TransactionModel(db.Model):
         return query
 
     @classmethod
-    def delete_transaction(cls, transaction_id: int):
-        cls.query.filter(cls.id == transaction_id).delete()
+    def delete_transaction(cls, transaction_id: int, user_id: int):
+        num_rows_deleted = cls.query.filter(cls.id == transaction_id, cls.user_id == user_id).delete()
         db.session.commit()
+        return num_rows_deleted
